64bit types functional, no longer has issues when importing both 32 and 64 bit qsorts
